Ch 14 Lab finished
Sorry it's all in one commit, I forgot to click the Commit to master button after I finished the first part -.-
For the first part: I edited the default.txt file to have all sort of pitfall whitespace characters to make sure everything worked fine. In fillDeafaultResponses() I added a boolean variable blankLine to keep track of if the previous line was blank or not to determine if the line needed to be appended, ignored, or start a new response.
Second part: Works almost exactly like part one. I moved all the keywords and their responses to the response.txt file along with, again, plenty of pitfalls to test. Keywords are delimited by commas, and keyword, resoponse pairs are seperated by "blank lines". In the fillResponseMap() I also employed the boolean variable blankLine. If the previous line was blank, then the current non-blank line contains keywords. I parsed the keywords by using the Sting.split() method and stored the splits in an arraylist called keywords. If the previous line is not blank than the line contains responses that need to be appened, either to the original empty string or to the previous line contents. Once we hit another blank line, the program maps all the keywords in the keywords arrayList to whatever is stored in response and then the arrayList is cleared and the response variable is reset to empty.
I used the String.trim() method to check if a line was blank and to make sure that the keywords to be mapped didn't contain any leading or trailing whitespace characters. I tried to make the comments as clear as possible.
Chapter 14 Lab is Complete and Final
I every much appreciated how sassy this support help desk is; gave me a giggle while coding :)